LeetCode GraphQL: print more default values
