+++ workspaces/leetcode-api/package.json
@@ -16,7 +16,7 @@
   "scripts": {
     "format": "prettier --color --write .",
     "lint": "eslint --color --max-warnings=0 .",
-    "scrape-graphql-schema": "ts-node src/scripts/scrapeGraphQLSchema.ts",
+    "scrape-graphql-schema": "ts-node src/scripts/scrape-graphql-schema/main.ts",
     "typecheck": "tsc --pretty --project .",
     "validate-graphql-schema": "ts-node src/scripts/validateGraphQLSchema.ts"
   },
