+++ workspaces/leetcode-api/src/scripts/scrape-graphql-schema/constants.ts
