+++ workspaces/leetcode-api/src/scripts/scrape-graphql-schema/main.ts
@@ -9,14 +9,14 @@ import { popMany } from "@code-chronicles/util/popMany";
 import { sleep } from "@code-chronicles/util/sleep";
 import { whileReturnsTrueAsync } from "@code-chronicles/util/whileReturnsTrueAsync";
 
-import { SCHEMA_FILE } from "../graphql-extraction/constants";
+import { SCHEMA_FILE } from "./constants";
 import {
   fetchGraphQLTypeInformation,
   type InnerType,
   type LeetCodeGraphQLType,
-} from "../fetchGraphQLTypeInformation";
-import { outputGraphQLSchema } from "../graphql-extraction/outputGraphQLSchema";
-import { readSeedGraphQLTypeNames } from "../graphql-extraction/readSeedGraphQLTypeNames";
+} from "../../fetchGraphQLTypeInformation";
+import { readSeedGraphQLTypeNames } from "./readSeedGraphQLTypeNames";
+import { stringifyGraphQLSchema } from "./stringifyGraphQLSchema";
 
 const BATCH_SIZE = 100;
 
@@ -96,7 +96,7 @@ async function main(): Promise<void> {
     return true;
   });
 
-  const schema = await outputGraphQLSchema(
+  const schema = await stringifyGraphQLSchema(
     [...typeInfos.values()].filter(isNonNullish),
   );
