+++ workspaces/leetcode-api/src/scripts/scrape-graphql-schema/readSeedGraphQLTypeNames.ts
